Change method to check if a pdf file has annotations.
The previous the method checking annotations of a pdf file does not work (at least for my laptop).
I change the method to that based on how many times the pdf file has been eddited.
After edditing pdf-files in Preview.app, Preview.app stamps "%%EOF" at the end of the pdf-file every saving.
The EOF stamp can be used for checking if the parsed pdf-file has been eddited & saved so far.
In the method I changed, the has_annotationss function counts the EOF-stamp, and if the number is higher than 2,
then the function returns true.